Add files via upload
As original code from CCERA is not working on actual versions of GNU Radio, i fixed some issues related to running a script of Python 2 in a Python 3 compiler of GNU Radio.
Mainly in Fb_helper, but i made some fixes in: .make, Folder_Block and  pulsar_filterbank_uhd.grc. More details in mods_for_python3_pulsar_filterbank (but for now in spanish).

I will explain more these fixes in future, but now i will explain how i make it work in bare minimum (still with issues in tset up and observe_pulsar.py):
 -You must clone main brach for Marcus and then you follow installation steps, until you will get an error in bash. Later you can replace original files with mine. 
-Open with GNU Radio the .grc file of mine. 
-16 bit mode output por . fil optional parameter is not working fine, same for  Roll-off correction (i have it on 0), if you have a ver low or high gain you can get a bash error. 

I´m using this code on my observations of pulsars at UPNA´s radiotelescope. This code generates a valid .fil archive that PRESTO can read. I think that i was able to see camelopardalis´s pulsar  or "B0329+54" according to my PRESTO outcomes. 

I at the time i made this fixes i didn´t know that someone tried before this, so mine fixes are from scratch and some help of chat GPT( I didn´t know how to use data in byte format for writing files in Python, one of key issues).

+++ Folder_Block.py
@@ -112,7 +112,7 @@ class blk(gr.sync_block):  # other base classes are basic_block, decim_block, in
             self.max_in = l
         elif (l < self.min_in):
             self.min_in = l
-        for i in range(l/self.flen):
+        for i in range(int(l/self.flen)):
             #
             # Do delay/dedispersion logic
             #
